Fix how blobs are converted.
Prior to this, blobs were being copied into their output buffer using
the full size of the output buffer instead of the size of the data
being copied.  This would result in reading off the data being copied.
Change this to the proper sizes, truncating if necessary.

This causes a regression test to need fixing too.